requirements.txt
openai
python-dotenv
moviepy
aiohttp
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
emoji
